Add files via upload
So here we have used Bar Chart to show top product categories with the highest number of ratings. It helps identify which product types are most popular or frequently reviewed by customers. Then we have used Clustered Bar Chart to show which categories tend to offer the highest or lowest average discounts. This can be useful for understanding pricing strategies across different product segments. We have used KPI as these are Simple card visuals to display single, important numbers at a glance.
We have also used Pie charts to show the sum of discount percentage with respect to review title and review content. Area chart is used for showing actual price and discounted percentage relationship. Line chart gives the counts of review content and sum of discount percentage with respect to product .